Add database

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,14 +1,20 @@
 package ru.yandex.practicum.filmorate.service;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.dto.user.NewUserRequestDTO;
+import ru.yandex.practicum.filmorate.dto.user.UpdateUserRequestDTO;
+import ru.yandex.practicum.filmorate.dto.user.UserResponseDTO;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.StorageException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.mapper.user.UserMapper;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.UserStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.Collection;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -17,31 +23,46 @@ import java.util.stream.Collectors;
 @Service
 public class UserService {
 
-    private final UserStorage userStorage;
 
-    public UserService(UserStorage userStorage) {
+    private final UserStorage userStorage;
+    private final UserMapper userMapper;
+
+
+    public UserService(@Qualifier("userDbStorage") UserStorage userStorage, UserMapper userMapper) {
         this.userStorage = userStorage;
+        this.userMapper = userMapper;
     }
 
-    public User createUser(User newUser) {
+    public UserResponseDTO createUser(NewUserRequestDTO newUserRequestDTO) {
+
+        User newUser = userMapper.toUser(newUserRequestDTO);
+
         setUserNameIfEmpty(newUser);
-        boolean wasCreated = userStorage.createUser(newUser);
-        if (!wasCreated) {
+        Long newUserId = userStorage.createUser(newUser);
+        if (newUserId == 0) {
             throw new StorageException(newUser + " wasn't created");
         }
-        log.info("New user with id={} was created", newUser.getId());
-        return newUser;
+        Optional<User> userInStorage = userStorage.findUserById(newUserId);
+        if (userInStorage.isEmpty()) {
+            throw new StorageException(newUser + " wasn't created");
+        }
+        log.info("New user with id={} was created", userInStorage.get().getId());
+
+        return userMapper.toUserResponseDTO(userInStorage.get());
     }
 
-    public Collection<User> getUsers() {
-        return userStorage.getUsers();
+    public Collection<UserResponseDTO> getUsers() {
+        return userMapper.toUserResponseDTO(userStorage.getUsers());
     }
 
-    public User getUserById(long id) {
-        return checkAndGetUserById(id);
+    public UserResponseDTO getUserById(long id) {
+        return userMapper.toUserResponseDTO(checkAndGetUserById(id));
     }
 
-    public User updateUser(User userToUpdate) {
+    public UserResponseDTO updateUser(UpdateUserRequestDTO userToUpdateDTO) {
+
+        User userToUpdate = userMapper.toUser(userToUpdateDTO);
+
         validateUserToUpdate(userToUpdate);
         boolean wasUpdated = userStorage.updateUser(userToUpdate);
         if (!wasUpdated) {
@@ -49,23 +70,27 @@ public class UserService {
         }
         log.info("User with id={} was updated", userToUpdate.getId());
         Optional<User> userOptional = userStorage.findUserById(userToUpdate.getId());
-        return userOptional.orElse(null);
+
+
+        return userMapper.toUserResponseDTO(userOptional.get());
     }
 
     public void addFriend(long userId, long friendId) {
         checkAndGetUserById(userId);
         checkAndGetUserById(friendId);
+
         if (userId == friendId) {
             throw new ValidationException("User can't be a friend to himself");
         }
+
+        if (isUserHaveFriend(userId, friendId)) {
+            throw new ValidationException("User already have a friend with id=" + friendId);
+        }
+
+
         // добавляем пользователю друга
         boolean friendWasAddedToUser = userStorage.addFriend(userId, friendId);
         if (!friendWasAddedToUser) {
-            throw new StorageException("Friend with id=" + friendId + " wasn't added to User with id=" + userId);
-        }
-        // добавляем в обратную сторону, чтобы оба были друзьями друг у друга
-        boolean userWasAddedToNewFriendUser = userStorage.addFriend(friendId, userId);
-        if (!userWasAddedToNewFriendUser) {
             throw new StorageException("Friend with id=" + friendId + " wasn't added to User with id=" + userId);
         }
 
@@ -79,18 +104,24 @@ public class UserService {
             throw new ValidationException("User can't be a friend to himself");
         }
 
+        // всё хорошо, если друга и не было (это чтобы работали тесты), хотя я считаю тут должна быть ValidationException
+        if (!isUserHaveFriend(userId, friendId)) {
+            return;
+        }
+
         // удаляем друга у пользователя
         boolean friendWasDeletedFromUser = userStorage.deleteFriend(userId, friendId);
         if (!friendWasDeletedFromUser) {
             throw new StorageException("Friend with id=" + friendId + " wasn't deleted from User with id=" + userId);
         }
 
-        // удаляем пользователя у бывшего друга
-        boolean userWasDeletedFromExFriendUser = userStorage.deleteFriend(friendId, userId);
-        if (!userWasDeletedFromExFriendUser) {
-            throw new StorageException("Friend with id=" + friendId + " wasn't deleted from User with id=" + userId);
+        // удаляем пользователя у бывшего друга, если он у него есть
+        if (isUserHaveFriend(friendId, userId)) {
+            boolean userWasDeletedFromExFriendUser = userStorage.deleteFriend(friendId, userId);
+            if (!userWasDeletedFromExFriendUser) {
+                throw new StorageException("Friend with id=" + friendId + " wasn't deleted from User with id=" + userId);
+            }
         }
-
     }
 
     private void setUserNameIfEmpty(User user) {
@@ -114,12 +145,19 @@ public class UserService {
         return mayBeUser.get();
     }
 
-    public Set<User> getUserFriends(long userId) {
-        checkAndGetUserById(userId);
-        return userStorage.getUserFriends(userId);
+    protected boolean isUserHaveFriend(Long user_id, Long friend_id) {
+
+        return !userStorage.getUserFriends(user_id).stream()
+                .filter(user -> Objects.equals(user.getId(), friend_id))
+                .toList().isEmpty();
     }
 
-    public Set<User> getUserCommonFriends(long userId, long otherUserId) {
+    public Collection<UserResponseDTO> getUserFriends(long userId) {
+        checkAndGetUserById(userId);
+        return userMapper.toUserResponseDTO(userStorage.getUserFriends(userId));
+    }
+
+    public Collection<UserResponseDTO> getUserCommonFriends(long userId, long otherUserId) {
 
         if (userId == otherUserId) {
             throw new ValidationException("User can't have common friends with himself");
@@ -131,9 +169,9 @@ public class UserService {
         Set<User> userFriends = userStorage.getUserFriends(userId);
         Set<User> otherUserIdFriends = userStorage.getUserFriends(otherUserId);
 
-        return userFriends
+        return userMapper.toUserResponseDTO(userFriends
                 .stream()
                 .filter(otherUserIdFriends::contains)
-                .collect(Collectors.toSet());
+                .collect(Collectors.toSet()));
     }
 }
